+++ paparazzi/paparazzi-annotation-processor/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=paparazzi-annotation-processor
+POM_NAME=Paparazzi Annotation Processor
+POM_DESCRIPTION=An annotation processor used to generate Paparazzi tests
+POM_PACKAGING=jar
