+++ paparazzi/paparazzi-annotation/build.gradle
