+++ paparazzi/paparazzi-annotation/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=paparazzi-annotation
+POM_NAME=Paparazzi Annotation API
+POM_DESCRIPTION=An API for annotating composable preview methods to generate Paparazzi tests
+POM_PACKAGING=jar
