+++ paparazzi/paparazzi-annotation/src/main/java/app/cash/paparazzi/annotation/api/ComposableWrapper.kt
@@ -1,0 +1,8 @@
+package app.cash.paparazzi.annotation.api
+
+import androidx.compose.runtime.Composable
+
+interface ComposableWrapper {
+  @Composable
+  fun wrap(content: @Composable () -> Unit)
+}
