+++ paparazzi/paparazzi-annotation/src/main/java/app/cash/paparazzi/annotation/api/types/NightMode.kt
@@ -1,0 +1,20 @@
+package app.cash.paparazzi.annotation.api.types
+
+import androidx.annotation.StringDef
+
+object NightMode {
+  const val DEFAULT = ""
+
+  const val NOTNIGHT = "NOTNIGHT"
+  const val NIGHT = "NIGHT"
+}
+
+@Retention(AnnotationRetention.SOURCE)
+@StringDef(
+  value = [
+    NightMode.DEFAULT,
+    NightMode.NOTNIGHT,
+    NightMode.NIGHT,
+  ]
+)
+annotation class DeviceNightMode
