+++ paparazzi/paparazzi-annotation/src/main/java/app/cash/paparazzi/annotation/api/types/ScreenOrientation.kt
@@ -1,0 +1,22 @@
+package app.cash.paparazzi.annotation.api.types
+
+import androidx.annotation.StringDef
+
+object ScreenOrientation {
+  const val DEFAULT = ""
+
+  const val PORTRAIT = "PORTRAIT"
+  const val LANDSCAPE = "LANDSCAPE"
+  const val SQUARE = "SQUARE"
+}
+
+@Retention(AnnotationRetention.SOURCE)
+@StringDef(
+  value = [
+    ScreenOrientation.DEFAULT,
+    ScreenOrientation.PORTRAIT,
+    ScreenOrientation.LANDSCAPE,
+    ScreenOrientation.SQUARE,
+  ]
+)
+annotation class DeviceOrientation
