--- paparazzi/paparazzi-api/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=paparazzi-api
-POM_NAME=Paparazzi API
-POM_DESCRIPTION=An API for annotating composable preview methods to generate Paparazzi tests
-POM_PACKAGING=jar
