--- paparazzi/paparazzi-api/src/main/java/app/cash/paparazzi/api/ComposableWrapper.kt
@@ -1,8 +1,0 @@
-package app.cash.paparazzi.api
-
-import androidx.compose.runtime.Composable
-
-interface ComposableWrapper {
-  @Composable
-  fun wrap(content: @Composable () -> Unit)
-}
