--- paparazzi/paparazzi-api/src/main/java/app/cash/paparazzi/api/types/NightMode.kt
@@ -1,20 +1,0 @@
-package app.cash.paparazzi.api.types
-
-import androidx.annotation.StringDef
-
-object NightMode {
-  const val DEFAULT = ""
-
-  const val NOTNIGHT = "NOTNIGHT"
-  const val NIGHT = "NIGHT"
-}
-
-@Retention(AnnotationRetention.SOURCE)
-@StringDef(
-  value = [
-    NightMode.DEFAULT,
-    NightMode.NOTNIGHT,
-    NightMode.NIGHT,
-  ]
-)
-annotation class DeviceNightMode
