--- paparazzi/paparazzi-api/src/main/java/app/cash/paparazzi/api/types/TouchScreen.kt
@@ -1,22 +1,0 @@
-package app.cash.paparazzi.api.types
-
-import androidx.annotation.StringDef
-
-object TouchScreen {
-  const val DEFAULT = ""
-
-  const val NOTOUCH = "NOTOUCH"
-  const val STYLUS = "STYLUS"
-  const val FINGER = "FINGER"
-}
-
-@Retention(AnnotationRetention.SOURCE)
-@StringDef(
-  value = [
-    TouchScreen.DEFAULT,
-    TouchScreen.NOTOUCH,
-    TouchScreen.STYLUS,
-    TouchScreen.FINGER,
-  ]
-)
-annotation class DeviceTouchScreen
