--- paparazzi/paparazzi-processor/gradle.properties
@@ -1,4 +1,0 @@
-POM_ARTIFACT_ID=paparazzi-processor
-POM_NAME=Paparazzi Annotation Processor
-POM_DESCRIPTION=An annotation processor used to generate Paparazzi tests
-POM_PACKAGING=jar
