+++ paparazzi/settings.gradle
@@ -2,9 +2,9 @@ rootProject.name = 'paparazzi-libs'
 
 include ':paparazzi'
 include ':paparazzi-agent'
-include ':paparazzi-api'
+include ':paparazzi-annotation'
+include ':paparazzi-annotation-processor'
 include ':paparazzi-gradle-plugin'
-include ':paparazzi-processor'
 include ':libs:layoutlib'
 include ':libs:native-macarm'
 include ':libs:native-macosx'
